// NenovCars.client-app/src/components/AdsContainer/interfaces/ICarAd.ts
export interface ICarAd {
    id: string;
    brand: string;
    imageUrl?: string;
    model: string;
    hp: string;
    fuel: string;
    adTitle: string;
    yearOfProduction: string;
    color: string;
    transmission: string;
    vehicleType: string;
    price: string;
    kilometre: string;
    description: string;
}